ellipsis-finance-generator-master/ellipsis-finance-generator-master/requirements.txt
reportlab
pypdf2
inflect
openpyxl
pymupdf
python-docx